first pass at package restructure
