+++ parliament_mcp/mcp_server/__init__.py
@@ -1,0 +1,1 @@
+"""MCP server for UK Parliament data."""
